+++ NEWS.md
@@ -1,6 +1,7 @@
 # deltaccd 0.2.0
 * Modified `calcCorr()`, `calcCCD()`, `calcDeltaCCD()` to use data.table.
 * Added optional `scale` argument to `calcCCD()` and `calcDeltaCCD()` to scale CCD by number of gene pairs.
+* Added reference correlations for human blood.
 
 # deltaccd 0.1.2
 * Fixed obvious bug in `plotHeatmap()`.